Recover controller command flow
Model cheat matcher results with a named enum and use the explicit no-command value where input is suppressed.

Recover the per-pattern history base so the matcher no longer needs an identical-arm branch to defeat CSE. Also give Think2contact the actual narrow GotoPosition prototype, which matches without the old type lie.

+++ src/main.exe/Think2contact.c
@@ -25,8 +25,7 @@
  * consumers use `Attrib` directly.
  */
 extern Humanoid *Me_THINK_C;
-/* Retail's own prototype drift (def: s16(s32, s32)) -- byte-required: correcting it changes the caller. */
-extern int GotoPosition(int vx, int vz);
+extern s16 GotoPosition(s32 vx, s32 vz);
 
 s16 Think2contact(void)
 {
